Restructure commands into dedicated directories with documentation
Reorganize CLI commands (init, whoami) into their own subdirectories under src/commands/<name>/index.ts instead of flat .ts files. This improves scalability and allows each command to have dedicated resources. Add comprehensive README documentation for all command groups (auth, config, env, init, whoami) with usage examples, options, and API endpoints. Update the deploy command README to clarify that it currently uses mocked data. All relative import paths updated accordingly.
